+++ src/feelpp/benchmarking/report/templates/benchmark.adoc.j2
@@ -17,9 +17,9 @@
 
 [%dynamic%close%hide_code,python]
 ----
-from feelpp.benchmarking.report.components.atomicReport import AtomicReportController
-from feelpp.benchmarking.report.components.atomicReport import AtomicReportModel
-from feelpp.benchmarking.report.components.atomicReport import AtomicReportView
+from feelpp.benchmarking.report.components.controllers import AtomicReportController
+from feelpp.benchmarking.report.components.models import AtomicReportModel
+from feelpp.benchmarking.report.components.views import AtomicReportView
 ----
 
 [%dynamic%close%hide_code,python]
