+++ src/feelpp/benchmarking/report/templates/machine.adoc.j2
@@ -1,0 +1,49 @@
+= Use case report
+:page-plotly: true
+:page-jupyter: true
+:page-tags: toolbox, catalog
+:parent-catalogs: {{parent_catalogs}}
+:description: 
+:page-illustration: 
+:revdate: 
+
+- Application : {{application_display_name}}
+- Machine : {{machine_display_name}}
+- Use case : {{use_case_display_name}}
+
+
+[%dynamic%close%hide_code,python]
+----
+from numpy import nan
+from feelpp.benchmarking.report.components.controllers import MachineController
+from feelpp.benchmarking.report.components.models import MachineModel
+from feelpp.benchmarking.report.components.views import MachineView
+----
+
+[%dynamic%close%hide_code,python]
+----
+model=MachineModel({{ reports_dfs }})
+view=MachineView({{ plots_config }})
+controller=MachineController(model,view)
+----
+
+[%dynamic%open%hide_code,python]
+----
+for fig in controller.generateAll():
+    fig.show()
+----
+
+++++
+<style>
+details>.title::before, details>.title::after {
+    visibility: hidden;
+}
+details>.content>.dynamic-py-result>.content>pre {
+    max-height: 100%;
+    padding: 0;
+    margin:16px;
+    background-color: white;
+    line-height:0;
+}
+</style>
+++++
